Simplifies build
* Moves styles.js & pdf.js into build.js

* Leverages mixins to make templates more DRY
