+++ package.json
@@ -9,11 +9,9 @@
   "scripts": {
     "start": "npm run browser_sync & NODE_ENV=development nodemon --exec 'npm run build && browser-sync reload'",
     "browser_sync": "browser-sync start --server _build --no-ui --no-notify",
-    "build:html": "node build.js",
-    "build:css": "node styles.js",
-    "build": "npm run build:html && npm run build:css",
+    "build": "node build.js",
     "build:minify": "NODE_ENV=production npm run build",
-    "pdf": "NODE_ENV=print npm run build && node pdf.js && npm run build",
+    "pdf": "NODE_ENV=print npm run build && npm run build",
     "deploy": "npm run build:minify && surge _build metalsmith-resume.lowmess.com"
   },
   "repository": {
