+++ templates/default.pug
@@ -4,38 +4,40 @@ include mixins
 head
   include partials/head
 
-body.bg-white.bg-light-gray-page.bg-white-p.sans-serif.dark-gray
-  article.mw-page.mv-5-page.center.ma0-p.shadow-2-page
+body.bg-white.bg-light-gray-page.bg-white-p.dark-gray(class= process.env.NODE_ENV == 'print' ? 'helvetica' : 'sans-serif')
+  article.mw-page.mv-5-page.center.ma0-p.shadow-2-page.lh-title
     include partials/header
 
     main.pa4.bg-white
       block content
 
         if site.jobs > 0
-          h3.f5.fw7.mt2.mb0.ttu: a.link.accent.underline-hover(href="/experience") Experience
-          section.work.cf.mb3
+          +section-title
+            +accent-link('Experience', '/experience')
+          +section
             - for (var i = 0; i < site.jobs; i++)
               +experience(collections.experience[i])
 
         if site.schools > 0
-          h3.f5.fw7.mt2.mb0.ttu: a.link.accent.underline-hover(href="/education") Education
-          section.education.cf.mb3
+          +section-title
+            +accent-link('Education', '/education')
+          +section
             - for (var i = 0; i < site.schools; i++)
               +experience(collections.education[i])
 
         if person.proficiencies.length > 0
-          section.skills.cf.mb3
-            .w-33-ns.w-33-p.fl-ns.fl-p
-              h3.f5.fw7.mt2.mb3.ttu.accent Proficiencies
-            ul.cc2-ns.cg3-ns.cc2-p.cg3-p.pl0.w-66-ns.w-66-p.fr-ns.fr-p.mt3.mt0-ns.mt0-p.lh-title.list-inside-ns
-              each skill in person.proficiencies
-                li= skill
+          +section
+            +left-column
+              +section-title('Proficiencies')
+            +right-column
+              ul.cc2-ns.cg3-ns.cc2-p.cg3-p.pl0.mt3.mt0-ns.mt0-p.list-inside-ns.list-inside-p
+                each skill in person.proficiencies
+                  li= skill
 
         if person.email && site.references
-          footer.mt4
-            hr.w4.ba.b--accent.bw1.ml0
-            p: a.link.dark-gray.underline-hover(href='mailto:' + person.email) References available upon request
+          +footer
+            p: +link('References available upon request', 'mailto:' + person.email)
 
   if site.promolink
     footer.pb4.tc.dn.dn-p.db-page
-      p.lh-title Made with <a href="https://github.com/lowmess/metalsmith-resume" class="accent link underline-hover">Metalsmith Resume</a> by <a href="https://github.com/lowmess/" class="accent link underline-hover">@lowmess</a>
+      p Made with #[+accent-link('Metalsmith Resume', 'https://github.com/lowmess/metalsmith-resume')] by #[+accent-link('@lowmess', 'https://github.com/lowmess/')]
