+++ templates/mixins.pug
@@ -1,31 +1,69 @@
+//- Left column of two column layout
+mixin left-column
+  .w-33-ns.w-33-p.fl-ns.fl-p.mt1.mb3.mb1-ns.mb1-p
+    block
+
+mixin right-column
+  .w-66-ns.w-66-p.fr-ns.fr-p.mv1
+    block
+
+//- Section elements
+mixin section-title(content)
+  h3.f5.fw7.mt2.mb3.ttu.accent
+    if block
+      block
+    else
+      = content
+
+mixin section
+  section.cf.mb4.mb3-ns.mb3-p
+    block
+
+//- Links
+mixin link(text, url)
+  a.link.dark-gray.underline-hover(href=url)= text
+
+mixin accent-link(text, url)
+  a.link.accent.underline-hover(href=url)= text
+
+mixin header-link(text, url)
+  a.link.bg-accent-text.underline-hover(href=url, class=site.printstyles ? 'dark-gray-p' : '')= text
+
+//- Footer
+mixin footer
+  footer.mt4
+    hr.w4.ba.b--accent.bw1.ml0
+    block
+
 //- Experiences are the molecular unit of the resume. They are what the markdown
     files in the source directory compile in to.
 mixin experience(exp)
-  section.experience.cf.mv3
+  +section
     //- Position
-    h3.f3.mv1.fw6= exp.title
+    h3.f3.mb2.mb0-p.fw6= exp.title
     //- Metadata
-    ul.w-33-ns.w-33-p.fl-ns.fl-p.list.ml0.pl0.lh-title.f5.mt1.fw2
-      //- Place of Work
-      li.fw5
-        if exp.url
-          a.link.dark-gray.underline-hover(href=exp.url)= exp.subtitle
-        else
-          = exp.subtitle
-      //- Time
-      li
-        time(datetime=moment.utc(exp.startDate))= moment.utc(exp.startDate).format("MMM YYYY")
-        | &#032;&ndash;&#032;
-        if exp.endDate
-          time(datetime=moment.utc(exp.endDate))= moment.utc(exp.endDate).format("MMM YYYY")
-        else
-          | Present
-      //- Location
-      if exp.location
-        li= exp.location
+    +left-column
+      ul.list.ml0.pl0.f5.mt1.fw2
+        //- Place of Work
+        li.fw5
+          if exp.url
+            +link(exp.subtitle, exp.url)
+          else
+            = exp.subtitle
+        //- Time
+        li
+          time(datetime=moment.utc(exp.startDate))= moment.utc(exp.startDate).format("MMM YYYY")
+          | &#032;&ndash;&#032;
+          if exp.endDate
+            time(datetime=moment.utc(exp.endDate))= moment.utc(exp.endDate).format("MMM YYYY")
+          else
+            | Present
+        //- Location
+        if exp.location
+          li= exp.location
 
     //- Contents
     if exp.contents
-      .w-66-ns.w-66-p.fr-ns.fr-p.mt1
-        .measure.lh-title.mt0.f5
+      +right-column
+        .measure.f5
           != exp.contents
